feat: add delete expense use case

+++ src/data/use-cases/delete-credit-card-use-case.ts
@@ -1,8 +1,10 @@
 import { DeleteCreditCard } from '@/domain/use-cases/delete-credit-card';
 import { UseCase } from '@/domain/use-cases/use-case';
+import { YouAreNotAllowedToChangeThisResourceError } from '@/domain/errors/you-no-have-permission-error';
 
 import { FindCreditCardByIdUseCase } from './find-credit-card-by-id-use-case';
 import { DeleteCreditCardRepository } from '../protocols/database/delete-credit-card-repository';
+import { FindUserByIdUseCase } from './find-user-by-id-use-case';
 
 export class DeleteCreditCardUseCase
   implements UseCase<DeleteCreditCard.Params, DeleteCreditCard.Result>
@@ -10,12 +12,21 @@ export class DeleteCreditCardUseCase
   constructor(
     private readonly findCreditCardByIdUseCase: FindCreditCardByIdUseCase,
     private readonly deleteCreditCardRepository: DeleteCreditCardRepository,
+    private readonly findUserByIdUseCase: FindUserByIdUseCase,
   ) {}
 
   async execute({
-    id,
+    creditCardId,
+    userId,
   }: DeleteCreditCard.Params): Promise<DeleteCreditCard.Result> {
-    const creditCard = await this.findCreditCardByIdUseCase.execute({ id });
+    const [user, creditCard] = await Promise.all([
+      this.findUserByIdUseCase.execute({ id: userId }),
+      this.findCreditCardByIdUseCase.execute({ id: creditCardId }),
+    ]);
+
+    if (creditCard.userId !== user.id) {
+      throw new YouAreNotAllowedToChangeThisResourceError();
+    }
 
     await this.deleteCreditCardRepository.delete({ id: creditCard.id });
   }
